-update Rigidbody Update 순서를 Graphic 컴포넌트 업데이트 이전으로 변경 Grid 수정 윈도우 크기 관련 에디터 설정 변경

+++ AmuEngine/AmuEngine/Sources/FrameCounter/FrameCounter.h
@@ -37,7 +37,7 @@ public:
         return deltaTime; 
     }
 
-    float getFrameCount() const
+    int getFrameCount() const
     {
         return frameCount;
     }
